Move TungsteniteConnector into Service module
