+++ src/service/maker.rs
@@ -63,8 +63,24 @@ where
 
 crate::cfg_feature! {
     #![feature = "tokio-tungstenite"]
-    use crate::transport::connector::TungsteniteConnector;
+
+    use crate::transport::TungsteniteApiTransport;
+    use futures_util::FutureExt;
+    use std::future::Future;
+    use std::pin::Pin;
     use tokio_tungstenite::tungstenite::client::IntoClientRequest;
+
+    /// A [`Service`] for creating new [`TungsteniteApiTransport`]s.
+    ///
+    /// This is used by [`tower::reconnect::Reconnect`] (used in
+    /// [`ClientBuilder`](crate::ClientBuilder)) for lazily connecting/reconnecting to websockets.
+    #[derive(Debug, Clone)]
+    pub struct TungsteniteConnector;
+}
+
+crate::cfg_feature! {
+    #![feature = "tokio-tungstenite"]
+    use crate::{Error, ErrorKind};
 
     impl<R> MakeApiService<TungsteniteConnector, R>
     where
@@ -73,6 +89,27 @@ crate::cfg_feature! {
         /// Creates a new [`MakeApiService`] using [`tokio_tungstenite`] as the underlying transport.
         pub fn new_tungstenite() -> Self {
             MakeApiService::new(TungsteniteConnector)
+        }
+    }
+
+    impl<R> Service<R> for TungsteniteConnector
+    where
+        R: IntoClientRequest + Unpin + Send + 'static,
+    {
+        type Response = TungsteniteApiTransport;
+        type Error = Error;
+        type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
+
+        fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+            Poll::Ready(Ok(()))
+        }
+
+        fn call(&mut self, request: R) -> Self::Future {
+            let transport = tokio_tungstenite::connect_async(request).map(|result| match result {
+                Ok((transport, _resp)) => Ok(TungsteniteApiTransport::new_tungstenite(transport)),
+                Err(e) => Err(Error::new(ErrorKind::ConnectionRefused).with_source(e)),
+            });
+            Box::pin(transport)
         }
     }
 }
